refactor/#77 : lib 폴더 내 유틸 파일들 중 공통이 아닌 도메인별로 사용되는 함수들을 각 도메인의 디렉토리로 분리하여 관리 구조를 명확하게 정리한다
- protest 관련 유틸 함수는 lib/heatmap.ts로 이동
- map 관련 유틸 함수는 lib/map.ts로 이동
- 공통 함수만 lib/utils.ts에 남김
- 각 도메인 내에서만 쓰는 함수가 전역에 위치하지 않도록 관심사에 따라 정리
